Refactored the HTML Parser (#3)
* feat: Improve the states

Moving away from the tokenizer dealing with script and style tags.

* feat: Added style and script handler

This way we can use DI to handle <style> and <script>, so we can use our
eventual CSS parser/JS parser

* fix: benches

* feat: Adding edition to workspace

* feat: Separated tokenizer

* feat: Moved depdencies to workspace

* refactor: Rename

* refactor: Moving away from concurrency

Specifically removing the Rc & Arc variants of dom nodes in favor of a
much simpler version

* refactor: Use new DOM structs

* docs: Updated the dom

* refactor: New HTML parser

Is now dependent on the consumer of the html parser to handle the
parsing loop, decoupling the responsibilities for the html parser to
focus on actually parsing.

* fix: benchmarks to reflect new changes

* fix: Processing inside style and script

* fix: Added async-trait to workspace

* delete: Parsing utility file

* feat: Event utility

* feat: Browser Core

Central orchestrator and controller for all browser functions, still
stuff to do

* patch: Network

Some adjustments

* feat: Added CLI

And adjusted the engine to use the `browser-core`

* refactor: Remove id to nodes

* feat: Allows printing body

* fix: Get rid of Collector Output

* delete: Event utility

It's harder to add browser events that depend on the TabId.

* feat: Added more tab details

* feat: More browser commands

* fix: Uses the TabId

* feat: UI update

Moving away from using iced in the renderer portion and am going to
create my own renderer probably using wgpu but for now basic navigation
and tab management is possible, no HTML anymore.

* delete: Event utility

* fix: Clippy hints